Use priv_dir/1 to get path to the priv directory
code:which(Module) searches the code path for the first file which
contains object code for Module and returns the absolute file name. When
used with embedded releases, this returns the name of the sendfile.ez
erlang archive which causes erl_ddll:load_driver/2 to fail.

Using code:priv_dir/1 fixes the problem.